Fling fixes and tests (#3191)
* begin fling tests

* fling fixes and tests

* Update test/move_effect_fling.c


* Update test/move_effect_fling.c


* Update test/move_effect_fling.c


* Update test/move_effect_fling.c


* Update test/move_effect_fling.c


* Update test/move_effect_fling.c


* Update test/move_effect_fling.c


* Update test/move_effect_fling.c


* scripts names change

* fix fling's test file

---------

+++ include/battle_util.h
@@ -190,7 +190,7 @@ bool32 SetIllusionMon(struct Pokemon *mon, u32 battlerId);
 bool8 ShouldGetStatBadgeBoost(u16 flagId, u8 battlerId);
 u8 GetBattleMoveSplit(u32 moveId);
 bool32 TestMoveFlags(u16 move, u32 flag);
-bool32 CanFling(u8 battlerId);
+bool32 CanFling(u32 battlerId);
 bool32 IsTelekinesisBannedSpecies(u16 species);
 bool32 IsHealBlockPreventingMove(u32 battler, u32 move);
 bool32 HasEnoughHpToEatBerry(u32 battlerId, u32 hpFraction, u32 itemId);
